refactor: update import links to components.
