+++ src/app/receipt/page.tsx
@@ -1,4 +1,4 @@
-import ReceiptPage from "../components/receiptPage";
+import ReceiptPage from "@/app/components/receipts/receiptPage";
 
 export default function Receipt() {
   return <ReceiptPage />;
